Move server to TypeScript
